+++ RELEASE_NOTES.rst
@@ -8,7 +8,7 @@ Release Notes
 ~~~~~
 
 - API: class :class:`.EstimatorDF` now implements the :class:`.HasExpressionRepr`
-  mixin, rendering estimator representations as :class:`.Expression` objects
+  mix-in, rendering estimator representations as :class:`.Expression` objects
   to enable better formatting